Added tests for auto sharding and moved all composition tests into a separate file, fixed a few bugs, some cleanup
