+++ machine_interface/src/composition.rs
@@ -14,12 +14,10 @@ use std::{
     vec,
 };
 
-#[cfg(test)]
-use crate::memory_domain::read_only::ReadOnlyContext;
-#[cfg(test)]
-use itertools::Itertools;
-
 mod join_iterator;
+
+#[cfg(test)]
+mod composition_tests;
 
 const OFFLOAD_CONST: usize = 2;
 
@@ -348,7 +346,6 @@ pub fn get_sharding(
     }
 
     // second create the iterators for all remaining shardings
-    // vec of (largest set size, max parallelism, chosen parallelism, processed)
     let mut any_set_groups = Vec::new();
     let mut join_iter = key_join_iter.map(|i| i as Box<dyn JoinIterator>);
     while i < join_order.len() {
@@ -359,29 +356,28 @@ pub fn get_sharding(
                     join_iter = join_iterator::SetAllIterator::new(join_iter, set, set_idx);
                 }
                 ShardingMode::Each => {
-                    let parallelism;
-                    (join_iter, parallelism) =
+                    let partitions;
+                    (join_iter, partitions) =
                         join_iterator::SetEachIterator::new(join_iter, set, set_idx);
-                    fixed_partitions *= parallelism;
+                    fixed_partitions *= partitions;
                 }
                 ShardingMode::AnyEach => {
-                    let set_size = set.size();
-                    let (any_join_iter, max_parallelism) = join_iterator::AnyIterator::new(
-                        join_iter,
-                        vec![set],
-                        vec![],
-                        vec![set_idx],
-                        sharding,
-                        min_set_size,
-                    );
-                    if max_parallelism > 0 {
-                        any_set_groups.push(AnySetGroup::new(set_size, max_parallelism));
+                    let (any_join_iter, largest_set_size, max_partitions) =
+                        join_iterator::AnyIterator::new(
+                            join_iter,
+                            vec![set],
+                            vec![],
+                            vec![set_idx],
+                            sharding,
+                            min_set_size,
+                        );
+                    if max_partitions > 0 {
+                        any_set_groups.push(AnySetGroup::new(largest_set_size, max_partitions));
                     }
                     join_iter = any_join_iter.map(|i| i as Box<dyn JoinIterator>);
                 }
                 ShardingMode::AnyKey => {
                     // get all sets that are joined together (i.e. find the next cross join)
-                    let mut largest_set_size = set.size();
                     let mut joined_sets = vec![set];
                     let mut joined_set_idcs = vec![set_idx];
                     let mut joined_strategies = vec![];
@@ -391,7 +387,6 @@ pub fn get_sharding(
                             let next_set_idx = join_order[i + 1];
                             // if the set is none we just skip it -> this allows for empty optional sets
                             if let Some((_, set)) = sets[next_set_idx].take() {
-                                largest_set_size = cmp::max(largest_set_size, set.size());
                                 joined_sets.push(set);
                                 joined_set_idcs.push(next_set_idx);
                                 joined_strategies.push(next_strategy);
@@ -403,16 +398,17 @@ pub fn get_sharding(
                         }
                     }
 
-                    let (any_join_iter, max_parallelism) = join_iterator::AnyIterator::new(
-                        join_iter,
-                        joined_sets,
-                        joined_strategies,
-                        joined_set_idcs,
-                        sharding,
-                        min_set_size,
-                    );
-                    if max_parallelism > 0 {
-                        any_set_groups.push(AnySetGroup::new(largest_set_size, max_parallelism));
+                    let (any_join_iter, largest_set_size, max_partitions) =
+                        join_iterator::AnyIterator::new(
+                            join_iter,
+                            joined_sets,
+                            joined_strategies,
+                            joined_set_idcs,
+                            sharding,
+                            min_set_size,
+                        );
+                    if max_partitions > 0 {
+                        any_set_groups.push(AnySetGroup::new(largest_set_size, max_partitions));
                     }
                     join_iter = any_join_iter.map(|i| i as Box<dyn JoinIterator>);
                 }
@@ -424,7 +420,7 @@ pub fn get_sharding(
         i += 1;
     }
 
-    // compute the target parallelism
+    // compute the target partitions
     let target_partitions = match any_sharding_mode {
         AnyShardingMode::MaxSharding => 0,
         AnyShardingMode::FixedSharding(n) => *n,
@@ -434,15 +430,11 @@ pub fn get_sharding(
             if total_largest_any_set_sizes == 0 {
                 0
             } else {
+                // use a minimal set size of at least 1 for this computation
+                let s_min = cmp::max(min_set_size, 1);
                 if total_largest_any_set_sizes
-                    > params.offload_const * min_set_size * params.num_local_cores
+                    > params.offload_const * s_min * params.num_local_cores
                 {
-                    log::trace!(
-                        "Any sets using at most local cores = {}",
-                        params.num_local_cores
-                    );
-                    params.num_local_cores
-                } else {
                     let all_cores =
                         params.num_local_cores + params.num_remote_cores.load(Ordering::Acquire);
                     log::trace!(
@@ -450,6 +442,12 @@ pub fn get_sharding(
                         all_cores
                     );
                     all_cores
+                } else {
+                    log::trace!(
+                        "Any sets using at most local cores = {}",
+                        params.num_local_cores
+                    );
+                    params.num_local_cores
                 }
             }
         }
@@ -461,17 +459,17 @@ pub fn get_sharding(
         target_partitions,
     );
 
-    // compute the parallelism for the any shardings
+    // compute the partitions for the any shardings
     if fixed_partitions < target_partitions && !any_set_groups.is_empty() {
-        let mut leftover_parallelism = target_partitions / fixed_partitions;
+        let mut leftover_partitions = target_partitions / fixed_partitions;
         loop {
             // find the best suitable set to parallelize over
             let mut best_idx = 0;
             let mut best_dist = 0.0;
             for (i, any_set_group) in any_set_groups.iter().enumerate() {
-                // check that we have not yet assigned a parallelism to this any set
+                // check that we have not yet assigned a partitions to this any set
                 if !any_set_group.processed {
-                    let remainder = any_set_group.max_partitions % leftover_parallelism;
+                    let remainder = any_set_group.max_partitions % leftover_partitions;
                     if remainder == 0 {
                         best_idx = i;
                         best_dist = 0.0;
@@ -490,27 +488,24 @@ pub fn get_sharding(
             if best_dist >= 1.0 {
                 any_set_groups[best_idx].target_partitions =
                     any_set_groups[best_idx].max_partitions;
-                leftover_parallelism /= any_set_groups[best_idx].max_partitions;
-                if leftover_parallelism <= 1 {
+                leftover_partitions /= any_set_groups[best_idx].max_partitions;
+                if leftover_partitions <= 1 {
                     break;
                 }
                 any_set_groups[best_idx].processed = true;
             } else {
                 if !any_set_groups[best_idx].processed {
-                    let chosen_parallelism = cmp::min(
-                        leftover_parallelism,
-                        any_set_groups[best_idx].max_partitions,
-                    );
-                    any_set_groups[best_idx].target_partitions = chosen_parallelism;
+                    any_set_groups[best_idx].target_partitions =
+                        cmp::min(leftover_partitions, any_set_groups[best_idx].max_partitions);
                 }
                 break;
             }
         }
     }
     if target_partitions > 0 {
-        trace!("Using any parallelism: {:?}", any_set_groups);
+        trace!("Using any partitions: {:?}", any_set_groups);
         if let Some(iter) = join_iter.as_mut() {
-            iter.reduce_any_parallelism(any_set_groups);
+            iter.reduce_any_partitions(any_set_groups);
         }
     }
 
@@ -530,784 +525,4 @@ pub fn get_sharding(
 
     trace!("Computed sharding: {:?}", final_sharding);
     final_sharding
-}
-
-//=======
-// TESTS
-
-/// Creates a dummy set from a vector of keys.
-/// The item indexes in the composition set are qual to the index of the key in the input.
-/// Keys do not need to be in correct order, but they will be sorted after producing.
-/// This is to allow to have keys with lower item indexes but higher keys and vice versa.
-#[cfg(test)]
-fn create_dummy_set(keys: Vec<u32>) -> CompositionSet {
-    let dummy_context: Arc<Context> = Arc::new(ReadOnlyContext::new_static::<u8>(&mut []));
-    let items = keys
-        .into_iter()
-        .enumerate()
-        .map(|(i, k)| {
-            (
-                DataItem {
-                    ident: i.to_string(),
-                    key: k,
-                    data: Position { offset: 0, size: 0 },
-                },
-                dummy_context.clone(),
-            )
-        })
-        .sorted_by_key(|tuple| tuple.0.key)
-        .collect();
-    CompositionSet {
-        item_list: items,
-        set_name: String::new(),
-    }
-}
-
-/// An array of options for expected sets in the input set vec produced by a sharing.
-#[cfg(test)]
-type SetGroup = Vec<Option<ExpectedSet>>;
-
-/// An array of tuples with the expected keys and item indexes for the items in a set.
-#[cfg(test)]
-type ExpectedSet = Vec<(u32, usize)>;
-
-#[cfg(test)]
-/// The expected is a list of all vectors of generated sets.
-fn check_sharding(actual: Vec<Vec<Option<CompositionSet>>>, expected: Vec<SetGroup>) {
-    assert_eq!(
-        expected.len(),
-        actual.len(),
-        "Not the number of set groups that were expected"
-    );
-    for (set_group_index, (actual_sets, expected_sets)) in
-        actual.into_iter().zip(expected.into_iter()).enumerate()
-    {
-        assert_eq!(
-            expected_sets.len(),
-            actual_sets.len(),
-            "Sets not matching for index {}, ",
-            set_group_index
-        );
-        for (set_index, (expected_set_opt, actual_set_opt)) in expected_sets
-            .into_iter()
-            .zip(actual_sets.into_iter())
-            .enumerate()
-        {
-            if expected_set_opt.is_none() {
-                assert!(
-                    actual_set_opt.is_none(),
-                    "Expexted none, but found a set for index {}",
-                    set_index
-                );
-                continue;
-            }
-            let mut expected_set = expected_set_opt.unwrap();
-            let mut actual_set = actual_set_opt.unwrap();
-            assert_eq!(expected_set.len(), actual_set.item_list.len());
-            // sort both lists by item index, since that one should be unique, since we only have a single context
-            expected_set.sort_by_key(|item| item.1.clone());
-            actual_set.item_list.sort_by_key(|item| item.0.key);
-            // have two sorted lists, check that each item index is the expected one and that it has the correct key
-            for ((expected_key, expected_ident), (actual_item, _)) in
-                expected_set.into_iter().zip(actual_set.item_list)
-            {
-                assert_eq!(
-                    expected_ident.to_string(),
-                    actual_item.ident,
-                    "for keys {}, {}",
-                    expected_key,
-                    actual_item.ident
-                );
-                assert_eq!(expected_key, actual_item.key);
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-fn print_sharding(actual: &Vec<Vec<Option<CompositionSet>>>) {
-    println!("Got sharding:");
-    for inv_sets in actual.iter() {
-        println!("[");
-        for (set_idx, set) in inv_sets.iter().enumerate() {
-            if set.is_none() {
-                println!("  set {}: [None]", set_idx);
-            } else {
-                print!("  set {}: [ ", set_idx);
-                for (item, _) in set.as_ref().unwrap().item_list.iter() {
-                    print!("{:?} ", item);
-                }
-                println!("]");
-            }
-        }
-        println!("]");
-    }
-}
-
-#[test]
-fn join_it_inner_test() {
-    let sets = vec![
-        Some((ShardingMode::Key, create_dummy_set(vec![3, 0, 1]))),
-        Some((ShardingMode::Key, create_dummy_set(vec![1, 0, 1, 2]))),
-    ];
-
-    let join_order = vec![0, 1];
-    let join_strategies = vec![JoinStrategy::Inner];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::MaxSharding,
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 1)]), Some(vec![(0, 1)])],
-        vec![Some(vec![(1, 2)]), Some(vec![(1, 0), (1, 2)])],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_left_test() {
-    let sets = vec![
-        Some((ShardingMode::Key, create_dummy_set(vec![0, 1, 2]))),
-        Some((ShardingMode::Key, create_dummy_set(vec![3, 1, 0, 1]))),
-    ];
-
-    let join_order = vec![0, 1];
-    let join_strategies = vec![JoinStrategy::Left];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::MaxSharding,
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 0)]), Some(vec![(0, 2)])],
-        vec![Some(vec![(1, 1)]), Some(vec![(1, 1), (1, 3)])],
-        vec![Some(vec![(2, 2)]), None],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_right_test() {
-    let sets = vec![
-        Some((ShardingMode::Key, create_dummy_set(vec![3, 1, 0, 1]))),
-        Some((ShardingMode::Key, create_dummy_set(vec![0, 1, 2]))),
-    ];
-
-    let join_order = vec![0, 1];
-    let join_strategies = vec![JoinStrategy::Right];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::MaxSharding,
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 2)]), Some(vec![(0, 0)])],
-        vec![Some(vec![(1, 1), (1, 3)]), Some(vec![(1, 1)])],
-        vec![None, Some(vec![(2, 2)])],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_outer_test() {
-    let sets = vec![
-        Some((ShardingMode::Key, create_dummy_set(vec![0, 1, 2]))),
-        Some((ShardingMode::Key, create_dummy_set(vec![3, 1, 0, 1]))),
-    ];
-
-    let join_order = vec![0, 1];
-    let join_strategies = vec![JoinStrategy::Outer];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::MaxSharding,
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 0)]), Some(vec![(0, 2)])],
-        vec![Some(vec![(1, 1)]), Some(vec![(1, 1), (1, 3)])],
-        vec![Some(vec![(2, 2)]), None],
-        vec![None, Some(vec![(3, 0)])],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_cross_test() {
-    let sets = vec![
-        Some((ShardingMode::Key, create_dummy_set(vec![0, 1, 2]))),
-        Some((ShardingMode::Key, create_dummy_set(vec![3, 1, 0, 1]))),
-    ];
-
-    let join_order = vec![0, 1];
-    let join_strategies = vec![JoinStrategy::Cross];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::MaxSharding,
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 0)]), Some(vec![(0, 2)])],
-        vec![Some(vec![(0, 0)]), Some(vec![(1, 1), (1, 3)])],
-        vec![Some(vec![(0, 0)]), Some(vec![(3, 0)])],
-        vec![Some(vec![(1, 1)]), Some(vec![(0, 2)])],
-        vec![Some(vec![(1, 1)]), Some(vec![(1, 1), (1, 3)])],
-        vec![Some(vec![(1, 1)]), Some(vec![(3, 0)])],
-        vec![Some(vec![(2, 2)]), Some(vec![(0, 2)])],
-        vec![Some(vec![(2, 2)]), Some(vec![(1, 1), (1, 3)])],
-        vec![Some(vec![(2, 2)]), Some(vec![(3, 0)])],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_order_test() {
-    let sets = vec![
-        Some((ShardingMode::Key, create_dummy_set(vec![3, 1, 0, 1]))),
-        Some((ShardingMode::Key, create_dummy_set(vec![0, 1, 2]))),
-    ];
-
-    let join_order = vec![1, 0];
-    let join_strategies = vec![JoinStrategy::Left];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::MaxSharding,
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 2)]), Some(vec![(0, 0)])],
-        vec![Some(vec![(1, 1), (1, 3)]), Some(vec![(1, 1)])],
-        vec![None, Some(vec![(2, 2)])],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_chain_test() {
-    let sets = vec![
-        Some((
-            ShardingMode::Key,
-            create_dummy_set(vec![1, 1234, 123, 124, 134]),
-        )),
-        Some((
-            ShardingMode::Key,
-            create_dummy_set(vec![2, 234, 1234, 123, 124]),
-        )),
-        Some((
-            ShardingMode::Key,
-            create_dummy_set(vec![3, 134, 234, 1234, 123]),
-        )),
-        Some((
-            ShardingMode::Key,
-            create_dummy_set(vec![4, 124, 134, 234, 1234]),
-        )),
-    ];
-
-    let join_order = vec![0, 1, 2, 3];
-    let join_strategies = vec![
-        JoinStrategy::Outer,
-        JoinStrategy::Outer,
-        JoinStrategy::Outer,
-    ];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::MaxSharding,
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(1, 0)]), None, None, None],
-        vec![None, Some(vec![(2, 0)]), None, None],
-        vec![None, None, Some(vec![(3, 0)]), None],
-        vec![None, None, None, Some(vec![(4, 0)])],
-        vec![
-            Some(vec![(123, 2)]),
-            Some(vec![(123, 3)]),
-            Some(vec![(123, 4)]),
-            None,
-        ],
-        vec![
-            Some(vec![(124, 3)]),
-            Some(vec![(124, 4)]),
-            None,
-            Some(vec![(124, 1)]),
-        ],
-        vec![
-            Some(vec![(134, 4)]),
-            None,
-            Some(vec![(134, 1)]),
-            Some(vec![(134, 2)]),
-        ],
-        vec![
-            None,
-            Some(vec![(234, 1)]),
-            Some(vec![(234, 2)]),
-            Some(vec![(234, 3)]),
-        ],
-        vec![
-            Some(vec![(1234, 1)]),
-            Some(vec![(1234, 2)]),
-            Some(vec![(1234, 3)]),
-            Some(vec![(1234, 4)]),
-        ],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_any_simple_test() {
-    let sets = vec![Some((
-        ShardingMode::AnyEach,
-        create_dummy_set(vec![0, 1, 2, 3]),
-    ))];
-
-    let join_order = vec![0];
-    let join_strategies = vec![];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::FixedSharding(2),
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 0), (1, 1)])],
-        vec![Some(vec![(2, 2), (3, 3)])],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_any_joined_keys_test() {
-    let sets = vec![
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 1, 2, 3, 5]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 2, 3, 4, 5]))),
-    ];
-
-    let join_order = vec![0, 1];
-    let join_strategies = vec![JoinStrategy::Inner];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::FixedSharding(2),
-        0,
-    );
-    let expected = vec![
-        vec![Some(vec![(0, 0), (2, 2)]), Some(vec![(0, 0), (2, 1)])],
-        vec![Some(vec![(3, 3), (5, 4)]), Some(vec![(3, 2), (5, 4)])],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_any_chain_test() {
-    let sets = vec![
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![0, 0, 0]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 1, 2, 3, 5]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 2, 3, 4, 5]))),
-    ];
-
-    let join_order = vec![1, 2, 0];
-    let join_strategies = vec![JoinStrategy::Inner, JoinStrategy::Cross];
-
-    let sharding = get_sharding(
-        sets,
-        join_order,
-        join_strategies,
-        &AnyShardingMode::FixedSharding(2),
-        0,
-    );
-    let expected = vec![
-        vec![
-            Some(vec![(0, 0), (0, 1), (0, 2)]),
-            Some(vec![(0, 0), (2, 2)]),
-            Some(vec![(0, 0), (2, 1)]),
-        ],
-        vec![
-            Some(vec![(0, 0), (0, 1), (0, 2)]),
-            Some(vec![(3, 3), (5, 4)]),
-            Some(vec![(3, 2), (5, 4)]),
-        ],
-    ];
-
-    print_sharding(&sharding);
-    check_sharding(sharding, expected);
-}
-
-#[test]
-fn join_it_any_chain_test_1() {
-    let sets = vec![
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![0, 1]))),
-        Some((
-            ShardingMode::AnyEach,
-            create_dummy_set(vec![0, 1, 2, 3, 4, 5]),
-        )),
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![0, 1, 2, 3]))),
-    ];
-
-    let join_order = vec![0, 1, 2];
-    let join_strategies = vec![JoinStrategy::Cross, JoinStrategy::Cross];
-
-    let sharding_6 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(6),
-        0,
-    );
-    assert_eq!(sharding_6.len(), 6); // <- should get 6 shardings
-    assert_eq!(sharding_6[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_6[0][0].as_ref().unwrap().len(), 2);
-    assert_eq!(sharding_6[0][1].as_ref().unwrap().len(), 1); // <- parallelized set
-    assert_eq!(sharding_6[0][2].as_ref().unwrap().len(), 4);
-
-    let sharding_3 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(3),
-        0,
-    );
-    assert_eq!(sharding_3.len(), 3); // <- should get 3 shardings
-    assert_eq!(sharding_3[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_3[0][0].as_ref().unwrap().len(), 2);
-    assert_eq!(sharding_3[0][1].as_ref().unwrap().len(), 2); // <- parallelized set
-    assert_eq!(sharding_3[0][2].as_ref().unwrap().len(), 4);
-
-    let sharding_4 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(4),
-        0,
-    );
-    assert_eq!(sharding_4.len(), 4); // <- should get 4 shardings
-    assert_eq!(sharding_4[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_4[0][0].as_ref().unwrap().len(), 2);
-    assert_eq!(sharding_4[0][1].as_ref().unwrap().len(), 6);
-    assert_eq!(sharding_4[0][2].as_ref().unwrap().len(), 1); // <- parallelized set
-}
-
-#[test]
-fn join_it_any_chain_test_2() {
-    let sets = vec![
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![0, 1, 2]))),
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![0, 1]))),
-    ];
-
-    let join_order = vec![0, 1];
-    let join_strategies = vec![JoinStrategy::Cross];
-
-    let sharding_2 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(2),
-        0,
-    );
-    assert_eq!(sharding_2.len(), 2); // <- should get 2 shardings
-    assert_eq!(sharding_2[0].len(), 2); // <- should still get 2 sets
-    assert_eq!(sharding_2[0][0].as_ref().unwrap().len(), 3);
-    assert_eq!(sharding_2[0][1].as_ref().unwrap().len(), 1); // <- parallelized set
-
-    let sharding_3 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(3),
-        0,
-    );
-    assert_eq!(sharding_3.len(), 3); // <- should get 3 shardings
-    assert_eq!(sharding_3[0].len(), 2); // <- should still get 2 sets
-    assert_eq!(sharding_3[0][0].as_ref().unwrap().len(), 1); // <- parallelized set
-    assert_eq!(sharding_3[0][1].as_ref().unwrap().len(), 2);
-
-    let sharding_6 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(6),
-        0,
-    );
-    assert_eq!(sharding_6.len(), 6); // <- should get 6 shardings
-    assert_eq!(sharding_6[0].len(), 2); // <- should still get 2 sets
-    assert_eq!(sharding_6[0][0].as_ref().unwrap().len(), 1); // <- parallelized set
-    assert_eq!(sharding_6[0][1].as_ref().unwrap().len(), 1); // <- parallelized set
-}
-
-#[test]
-fn join_it_any_chain_test_3() {
-    let sets = vec![
-        Some((ShardingMode::Each, create_dummy_set(vec![0, 0]))),
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![0, 0, 0]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 1, 1]))),
-    ];
-
-    let join_order = vec![0, 1, 2];
-    let join_strategies = vec![JoinStrategy::Cross, JoinStrategy::Cross];
-
-    let sharding_2 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(2),
-        0,
-    );
-    assert_eq!(sharding_2.len(), 2); // <- should get 2 shardings
-    assert_eq!(sharding_2[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_2[0][0].as_ref().unwrap().len(), 1); // <- always parallelized (not any)
-    assert_eq!(sharding_2[0][1].as_ref().unwrap().len(), 3);
-    assert_eq!(sharding_2[0][2].as_ref().unwrap().len(), 3);
-
-    let sharding_4 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(4),
-        0,
-    );
-    assert_eq!(sharding_4.len(), 4); // <- should get 4 shardings
-    assert_eq!(sharding_4[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_4[0][0].as_ref().unwrap().len(), 1); // <- always parallelized (not any)
-    assert_eq!(sharding_4[0][1].as_ref().unwrap().len(), 3);
-    assert_eq!(sharding_4[0][2].as_ref().unwrap().len(), 1); // <- parallelized any set
-
-    let sharding_6 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(6),
-        0,
-    );
-    assert_eq!(sharding_6.len(), 6); // <- should get 6 shardings
-    assert_eq!(sharding_6[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_6[0][0].as_ref().unwrap().len(), 1); // <- always parallelized (not any)
-    assert_eq!(sharding_6[0][1].as_ref().unwrap().len(), 1); // <- parallelized any set
-    assert_eq!(sharding_6[0][2].as_ref().unwrap().len(), 3);
-}
-
-#[test]
-fn join_it_any_chain_test_4() {
-    let sets = vec![
-        Some((ShardingMode::Each, create_dummy_set(vec![0, 0]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 2, 0, 1, 3]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 1, 1, 4]))),
-    ];
-
-    let join_order = vec![1, 2, 0];
-
-    let join_strategies_inner = vec![JoinStrategy::Inner, JoinStrategy::Cross];
-    let sharding_inner = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies_inner.clone(),
-        &AnyShardingMode::FixedSharding(10),
-        0,
-    );
-    assert_eq!(sharding_inner.len(), 4); // <- should get 4 shardings
-    assert_eq!(sharding_inner[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_inner[0][0].as_ref().unwrap().len(), 1); // <- always parallelized (not any)
-    assert_eq!(sharding_inner[0][1].as_ref().unwrap().len(), 2); // <- parallelized any set
-    assert_eq!(sharding_inner[0][2].as_ref().unwrap().len(), 1); // <- parallelized any set
-
-    let join_strategies_left = vec![JoinStrategy::Left, JoinStrategy::Cross];
-    let sharding_left = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies_left.clone(),
-        &AnyShardingMode::FixedSharding(10),
-        0,
-    );
-    assert_eq!(sharding_left.len(), 8); // <- should get 2*4=8 shardings
-    assert_eq!(sharding_left[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_left[0][0].as_ref().unwrap().len(), 1); // <- always parallelized (not any)
-    assert_eq!(sharding_left[0][1].as_ref().unwrap().len(), 2); // <- parallelized any set
-    assert_eq!(sharding_left[0][2].as_ref().unwrap().len(), 1); // <- parallelized any set
-
-    let join_strategies_right = vec![JoinStrategy::Right, JoinStrategy::Cross];
-    let sharding_right = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies_right.clone(),
-        &AnyShardingMode::FixedSharding(10),
-        0,
-    );
-    assert_eq!(sharding_right.len(), 6); // <- should get 2*3=6 shardings
-    assert_eq!(sharding_right[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_right[0][0].as_ref().unwrap().len(), 1); // <- always parallelized (not any)
-    assert_eq!(sharding_right[0][1].as_ref().unwrap().len(), 2); // <- parallelized any set
-    assert_eq!(sharding_right[0][2].as_ref().unwrap().len(), 1); // <- parallelized any set
-
-    let join_strategies_outer = vec![JoinStrategy::Outer, JoinStrategy::Cross];
-    let sharding_outer = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies_outer.clone(),
-        &AnyShardingMode::FixedSharding(10),
-        0,
-    );
-    assert_eq!(sharding_outer.len(), 10); // <- should get 2*5=10 shardings
-    assert_eq!(sharding_outer[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding_outer[0][0].as_ref().unwrap().len(), 1); // <- always parallelized (not any)
-    assert_eq!(sharding_outer[0][1].as_ref().unwrap().len(), 2); // <- parallelized any set
-    assert_eq!(sharding_outer[0][2].as_ref().unwrap().len(), 1); // <- parallelized any set
-}
-
-#[test]
-fn join_it_any_chain_test_5() {
-    let sets = vec![Some((
-        ShardingMode::AnyKey,
-        create_dummy_set(vec![0, 2, 0, 1, 3]),
-    ))];
-
-    let join_order = vec![0];
-    let join_strategies = vec![];
-
-    let sharding_2 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(2),
-        0,
-    );
-    assert_eq!(sharding_2.len(), 2); // <- should get 2 shardings
-    assert_eq!(sharding_2[0].len(), 1); // <- should still get 1 set
-    assert_eq!(sharding_2[0][0].as_ref().unwrap().len(), 3); // <- parallelized any set
-
-    let sharding_4 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(4),
-        0,
-    );
-    assert_eq!(sharding_4.len(), 4); // <- should get 4 shardings
-    assert_eq!(sharding_4[0].len(), 1); // <- should still get 1 set
-    assert_eq!(sharding_4[0][0].as_ref().unwrap().len(), 2); // <- parallelized any set
-}
-
-#[test]
-fn join_it_any_chain_test_6() {
-    let sets = vec![
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 1, 4]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 2, 0, 3]))),
-        Some((ShardingMode::AnyKey, create_dummy_set(vec![0, 1, 1]))),
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![5, 5]))),
-    ];
-
-    let join_order = vec![1, 2, 0, 3];
-    let join_strategies = vec![
-        JoinStrategy::Inner,
-        JoinStrategy::Right,
-        JoinStrategy::Cross,
-    ];
-
-    let sharding_2 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(2),
-        0,
-    );
-    assert_eq!(sharding_2.len(), 2); // <- should get 2 shardings
-    assert_eq!(sharding_2[0].len(), 4); // <- should still get 4 sets
-    assert_eq!(sharding_2[0][0].as_ref().unwrap().len(), 3);
-    assert_eq!(sharding_2[0][1].as_ref().unwrap().len(), 2);
-    assert_eq!(sharding_2[0][2].as_ref().unwrap().len(), 1);
-    assert_eq!(sharding_2[0][3].as_ref().unwrap().len(), 1); // <- parallelized any set
-    assert_eq!(sharding_2[0][0].as_ref().unwrap().item_list[0].0.key, 0);
-    assert_eq!(sharding_2[0][0].as_ref().unwrap().item_list[1].0.key, 1);
-    assert_eq!(sharding_2[0][0].as_ref().unwrap().item_list[2].0.key, 4);
-    assert_eq!(sharding_2[0][1].as_ref().unwrap().item_list[0].0.key, 0);
-    assert_eq!(sharding_2[0][1].as_ref().unwrap().item_list[1].0.key, 0);
-    assert_eq!(sharding_2[0][2].as_ref().unwrap().item_list[0].0.key, 0);
-    assert_eq!(sharding_2[0][3].as_ref().unwrap().item_list[0].0.key, 5);
-
-    let sharding_3 = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(3),
-        0,
-    );
-    assert_eq!(sharding_3.len(), 3); // <- should get 3 shardings
-    assert_eq!(sharding_3[0].len(), 4); // <- should still get 4 sets
-    assert_eq!(sharding_3[0][0].as_ref().unwrap().len(), 1); // <- parallelized any set
-    assert_eq!(sharding_3[0][1].as_ref().unwrap().len(), 2); // <- parallelized any set
-    assert_eq!(sharding_3[0][2].as_ref().unwrap().len(), 1); // <- parallelized any set
-    assert_eq!(sharding_3[0][3].as_ref().unwrap().len(), 2);
-    assert_eq!(sharding_3[0][0].as_ref().unwrap().item_list[0].0.key, 0);
-    assert_eq!(sharding_3[0][1].as_ref().unwrap().item_list[0].0.key, 0);
-    assert_eq!(sharding_3[0][1].as_ref().unwrap().item_list[1].0.key, 0);
-    assert_eq!(sharding_3[0][2].as_ref().unwrap().item_list[0].0.key, 0);
-    assert_eq!(sharding_3[0][3].as_ref().unwrap().item_list[0].0.key, 5);
-    assert_eq!(sharding_3[0][3].as_ref().unwrap().item_list[1].0.key, 5);
-}
-
-#[test]
-fn join_it_any_chain_test_7() {
-    let sets = vec![
-        Some((ShardingMode::Each, create_dummy_set(vec![0]))),
-        Some((ShardingMode::AnyEach, create_dummy_set(vec![0]))),
-        None,
-    ];
-
-    let join_order = vec![0, 1, 2];
-    let join_strategies = vec![JoinStrategy::Cross, JoinStrategy::Cross];
-
-    let sharding = get_sharding(
-        sets.clone(),
-        join_order.clone(),
-        join_strategies.clone(),
-        &AnyShardingMode::FixedSharding(usize::MAX),
-        0,
-    );
-    assert_eq!(sharding.len(), 1); // <- should get 1 sharding
-    assert_eq!(sharding[0].len(), 3); // <- should still get 3 sets
-    assert_eq!(sharding[0][0].as_ref().unwrap().len(), 1);
-    assert_eq!(sharding[0][1].as_ref().unwrap().len(), 1);
-    assert!(sharding[0][2].is_none());
-    assert_eq!(sharding[0][0].as_ref().unwrap().item_list[0].0.key, 0);
-    assert_eq!(sharding[0][1].as_ref().unwrap().item_list[0].0.key, 0);
 }
